cli: moved command.go to its own package to be reused
