+++ cli/borges/init.go
@@ -3,6 +3,7 @@ package main
 import (
 	"fmt"
 
+	bcli "github.com/src-d/borges/cli"
 	"gopkg.in/src-d/core-retrieval.v0/schema"
 	"gopkg.in/src-d/go-cli.v0"
 	"gopkg.in/src-d/go-log.v1"
@@ -14,11 +15,11 @@ func init() {
 
 type initCmd struct {
 	cli.Command `name:"init" short-description:"initialize the database schema" long-description:"Connects to the database and initializes the schema."`
-	databaseOpts
+	bcli.DatabaseOpts
 }
 
 func (c *initCmd) Execute(args []string) error {
-	db, err := c.openDatabase()
+	db, err := c.OpenDatabase()
 	if err != nil {
 		return fmt.Errorf("unable to get database: %s", err)
 	}
